ShippingSystem/requirements.txt
flask==3.1.0
bcrypt==4.2.1
cryptography==43.0.3
apscheduler==3.10.4
requests==2.32.3
usaddress==0.5.10
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
